+++ src/Paket.Bootstrapper/Properties/AssemblyInfo.cs
@@ -4,10 +4,11 @@ using System.Reflection;
 [assembly: AssemblyTitleAttribute("Paket.Bootstrapper")]
 [assembly: AssemblyProductAttribute("Paket")]
 [assembly: AssemblyDescriptionAttribute("A package dependency manager for .NET with support for NuGet packages and GitHub repositories.")]
-[assembly: AssemblyVersionAttribute("0.29.1")]
-[assembly: AssemblyFileVersionAttribute("0.29.1")]
+[assembly: AssemblyVersionAttribute("0.30.0")]
+[assembly: AssemblyFileVersionAttribute("0.30.0")]
+[assembly: AssemblyInformationalVersionAttribute("0.30.0-alpha009")]
 namespace System {
     internal static class AssemblyVersionInformation {
-        internal const string Version = "0.29.1";
+        internal const string Version = "0.30.0";
     }
 }