+++ src/Paket/Commands.fs
@@ -15,6 +15,8 @@ type Command =
     | [<First>][<CustomCommandLine("restore")>]             Restore
     | [<First>][<CustomCommandLine("simplify")>]            Simplify
     | [<First>][<CustomCommandLine("update")>]              Update
+    | [<First>][<CustomCommandLine("pack")>]                Pack
+    | [<First>][<CustomCommandLine("push")>]                Push
 with 
     interface IArgParserTemplate with
         member __.Usage = ""
@@ -118,5 +120,20 @@ type UpdateArgs =
     | Hard
     | Redirects
 with 
+    interface IArgParserTemplate with
+        member __.Usage = ""
+
+type PackArgs =
+    | [<CustomCommandLine("output")>][<Mandatory>] Output of string
+    | [<CustomCommandLine("buildconfig")>] BuildConfig of string
+with
+    interface IArgParserTemplate with
+        member __.Usage = ""
+
+type PushArgs =
+    | [<CustomCommandLine("url")>][<Mandatory>] Url of string
+    | [<CustomCommandLine("packagedir")>][<Mandatory>] PackageDir of string
+    | [<CustomCommandLine("apikey")>] ApiKey of string
+with
     interface IArgParserTemplate with
         member __.Usage = ""