+++ RELEASE_NOTES.md
@@ -1,4 +1,6 @@
-#### 0.29.1 - 19.02.2015
+#### 0.30.0-alpha010 - 19.02.2015
+* New command: `paket pack` - http://fsprojects.github.io/Paket/paket-pack.html
+* New command: `paket push` - http://fsprojects.github.io/Paket/paket-push.html
 * BUGFIX: fix no_auto_restore option parsing  - https://github.com/fsprojects/Paket/issues/632
 
 #### 0.29.0 - 18.02.2015